Finished this MCVE.
It turns out that imgui handles a button release followed by a mouse
move outside of the window correctly. During my tests I probably
moved my mouse outside of the button before releasing it.

I haven't tested what happens when include a loss of focus instead
however -- it is still possible that if you release the mouse
on the button, then move to another window and click the mouse
button there, causing the previous window to lose focus, all before
the next frame starts, that this button release is lost.